feat/HC-110 채널 회원 crud

+++ src/main/java/com/example/coconote/api/section/service/SectionService.java
@@ -5,8 +5,9 @@ import com.example.coconote.api.section.dto.request.SectionUpdateReqDto;
 import com.example.coconote.api.section.dto.response.SectionListResDto;
 import com.example.coconote.api.section.entity.Section;
 import com.example.coconote.api.section.repository.SectionRepository;
-import com.example.coconote.api.workspace.entity.Workspace;
-import com.example.coconote.api.workspace.repository.WorkspaceRepository;
+import com.example.coconote.api.workspace.workspace.entity.Workspace;
+import com.example.coconote.api.workspace.workspace.repository.WorkspaceRepository;
+import com.example.coconote.common.IsDeleted;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -29,18 +30,22 @@ public class SectionService {
 
     public SectionListResDto sectionCreate(SectionCreateReqDto dto) {
         Workspace workspace = workspaceRepository.findById(dto.getWorkspaceId()).orElseThrow(()-> new EntityNotFoundException("존재하지 않는 워크스페이스입니다."));
-
+        if(workspace.getIsDeleted().equals(IsDeleted.Y)) {
+            throw new IllegalArgumentException("이미 삭제된 워크스페이스입니다.");
+        }
         Section section = dto.toEntity(workspace);
         sectionRepository.save(section);
 
-        SectionListResDto resDto = section.fromEntity();
-
-        return resDto;
+        return section.fromEntity();
     }
 
 
-    public List<SectionListResDto> sectionList() {
-        List<Section> sections = sectionRepository.findAll();
+    public List<SectionListResDto> sectionList(Long workspaceId) {
+        Workspace workspace = workspaceRepository.findById(workspaceId).orElseThrow(()-> new EntityNotFoundException("워크스페이스를 찾을 수 없습니다."));
+        if(workspace.getIsDeleted().equals(IsDeleted.Y)) {
+            throw new IllegalArgumentException("이미 삭제된 워크스페이스입니다.");
+        }
+        List<Section> sections = sectionRepository.findByWorkspaceAndIsDeleted(workspace, IsDeleted.N);
         List<SectionListResDto> dtos = new ArrayList<>();
         for(Section s : sections) {
             dtos.add(s.fromEntity());
@@ -49,7 +54,10 @@ public class SectionService {
     }
 
     public SectionListResDto sectionUpdate(Long id, SectionUpdateReqDto dto) {
-        Section section = sectionRepository.findById(id).orElseThrow(()->new EntityNotFoundException("section not found"));
+        Section section = sectionRepository.findById(id).orElseThrow(()->new EntityNotFoundException("섹션을 찾을 수 없습니다."));
+        if(section.getIsDeleted().equals(IsDeleted.Y)) {
+            throw new IllegalArgumentException("이미 삭제된 섹션입니다.");
+        }
         section.updateEntity(dto);
         SectionListResDto resDto = section.fromEntity();
         return resDto;
@@ -57,9 +65,10 @@ public class SectionService {
 
 
     public void sectionDelete(Long id) {
-        Section section = sectionRepository.findById(id).orElseThrow(()->new EntityNotFoundException("section not found"));
-        Workspace workspace = workspaceRepository.findById(section.getWorkspace().getWorkspaceId()).orElse(null);
+        Section section = sectionRepository.findById(id).orElseThrow(()->new EntityNotFoundException("섹션을 찾을 수 없습니다."));
+        if(section.getIsDeleted().equals(IsDeleted.Y)) {
+            throw new IllegalArgumentException("이미 삭제된 섹션입니다.");
+        }
         section.deleteEntity();
-        workspaceRepository.save(workspace);
     }
 }
